regex ajoutée pour empecher l'injection de code et sécuriser les mots de passe

+++ backend/controllers/sauces.js
@@ -2,19 +2,21 @@ const Sauce = require('../models/Sauce');
 const fs = require('fs');
 
 exports.createSauce = (req, res, next) => {
-    const thingObject = JSON.parse(req.body.sauce);
-    delete thingObject._id;
-    const sauce = new Sauce({
-        ...thingObject,
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-        likes: 0,
-        dislikes:0,
-        usersLiked: [],
-        usersDisliked: []
-    });
-    sauce.save()
-        .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
-        .catch(error => res.status(500).json({ error }));
+    if (verifyInput(req)) {
+        const thingObject = JSON.parse(req.body.sauce);
+        delete thingObject._id;
+        const sauce = new Sauce({
+            ...thingObject,
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+            likes: 0,
+            dislikes:0,
+            usersLiked: [],
+            usersDisliked: []
+        });
+        sauce.save()
+            .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
+            .catch(error => res.status(500).json({ error }));
+    }
 };
 
 exports.getOneSauce = (req, res, next) => {
@@ -24,22 +26,24 @@ exports.getOneSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.findOne({ _id: req.params.id })
-        .then(sauce => {
-            if (req.file) {
-                const filename = sauce.imageUrl.split('/images/')[1];
-                fs.unlink(`images/${filename}`, () => { });
-            }
-            const thingObject = req.file ?
-            {
-                ...JSON.parse(req.body.sauce),
-                imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-            } : { ...req.body };
-            Sauce.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
-                .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
-                .catch(error => res.status(400).json({ error }));
-        })
-        .catch(error => res.status(501).json({ error }));
+    if (verifyInput(req)) {
+        Sauce.findOne({ _id: req.params.id })
+            .then(sauce => {
+                if (req.file) {
+                    const filename = sauce.imageUrl.split('/images/')[1];
+                    fs.unlink(`images/${filename}`, () => { });
+                }
+                const thingObject = req.file ?
+                {
+                    ...JSON.parse(req.body.sauce),
+                    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                } : { ...req.body };
+                Sauce.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
+                    .catch(error => res.status(400).json({ error }));
+            })
+            .catch(error => res.status(501).json({ error }));
+    }
     
 }
 
@@ -61,3 +65,17 @@ exports.getAllSauces = (req, res, next) => {
         .then(sauces => res.status(200).json(sauces))
         .catch(error => res.status(400).json({ error }));
 };
+
+function verifyInput (req) {
+    const test = /[A-Za-z éèçàêëñöùä\-]/;
+    if (
+        test.test(req.body.name) &&
+        test.test(req.body.manufacturer) &&
+        test.test(req.body.description) &&
+        test.test(req.body.mainPepper) 
+    ) {
+        return true;
+    } else {
+        return false;
+    }
+}
